config: allow mesh mode and user_mpm options
Modify network mode to support mode number 5 when CONFIG_MESH is
enabled.

Add user_mpm config parameter, when this is set to 1 (the default) the
peer link management is done on userspace, otherwise the peer management
will be done by the kernel.

Signed-off-by: Javier Lopez <jlopex@cozybit.com>
Signed-off-by: Jason Mobarak <jam@cozybit.com>

---
v4: Reordered to earlier in patch series
v5: Make user_mpm=1 the default